pyproject.toml
[project]
name = "ocr-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "fastapi>=0.116.1",
    "google-cloud-vision>=3.10.2",
    "numpy>=2.2.6",
    "openai>=1.97.0",
    "opencv-python>=4.12.0.88",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "simplejson>=3.20.1",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
]
